+++ km3pipe/io/aanet.py
@@ -135,13 +135,13 @@ class AanetPump(Pump):
                 #livetime_err = self.header.get_field('livetime', 1)
                 self.livetime_err = float(livetime_err)
                 self.livetime = float(livetime)
-            except (ValueError, UnicodeEncodeError):
+            except (ValueError, UnicodeEncodeError, AttributeError):
                 log.warn(filename + ": can't read livetime.")
                 self.livetime = 0
             try:
                 ngen = self.header.get_field('genvol', 4)
                 self.ngen = float(ngen)
-            except (ValueError, UnicodeEncodeError):
+            except (ValueError, UnicodeEncodeError, AttributeError):
                 log.warn(filename + ": can't read ngen.")
                 self.ngen = 0
             try: