+++ km3pipe/io/hdf5.py
@@ -18,7 +18,7 @@ import km3pipe as kp
 from km3pipe.core import Pump, Module, Blob
 from km3pipe.dataclasses import KM3Array, deserialise_map
 from km3pipe.logger import logging
-from km3pipe.tools import camelise, decamelise, split
+from km3pipe.dev import camelise, decamelise, split
 
 log = logging.getLogger(__name__)  # pylint: disable=C0103
 
@@ -59,7 +59,7 @@ class HDF5Sink(Module):
         # might be able to set to `None`, I don't know...
         self.n_rows_expected = self.get('n_rows_expected') or 10000
 
-        self.index = 1
+        self.index = 0
         self.h5file = tb.open_file(self.filename, mode="w", title="KM3NeT")
         try:
             self.filters = tb.Filters(complevel=5, shuffle=True,
@@ -186,7 +186,7 @@ class HDF5Pump(Pump):
             if os.path.isfile(fn):
                 h5file = tb.open_file(fn, 'r')
                 if not self.skip_version_check:
-                    self._check_version(fn)
+                    self._check_version(h5file, fn)
             else:
                 raise IOError("No such file or directory: '{0}'"
                               .format(fn))
@@ -210,9 +210,9 @@ class HDF5Pump(Pump):
         self.index = None
         self._reset_index()
 
-    def _check_version(self, filename):
+    def _check_version(self, h5file, filename):
         try:
-            version = np.string_(self.h5file.root._v_attrs.format_version)
+            version = np.string_(h5file.root._v_attrs.format_version)
         except AttributeError:
             log.error("Could not determine HDF5 format version: '%s'."
                       "You may encounter unexpected errors! Good luck..."