+++ km3pipe/io/root.py
@@ -22,7 +22,8 @@ import km3pipe as kp
 from km3pipe.core import Pump, Module
 from km3pipe.dataclasses import KM3Array, deserialise_map
 from km3pipe.logger import logging
-from km3pipe.tools import camelise, decamelise, insert_prefix_to_dtype, split
+from km3pipe.dev import camelise, decamelise, split
+from km3pipe.tools import insert_prefix_to_dtype
 
 log = logging.getLogger(__name__)  # pylint: disable=C0103